+++ src/app/garfile/amg/checkin/index.js
@@ -1,5 +1,4 @@
 const getController = require('./get.controller');
-const postController = require('./post.controller');
 const nav = require('../../../../common/utils/nav');
 
-module.exports = nav.buildRouterAndPaths('/garfile/amg/checkin', getController, postController);
+module.exports = nav.buildRouterAndPaths('/garfile/amg/checkin', getController);
